Actually, we can just look in the text for "[ ]"
Works surprisingly well, but it does take a long time. Have to hackly override
some pdfminer classes, so we just completly skip any character that's not a [ ],
(because splitting everything into individual characters adds like 20 seconds,
so we avoid that), and then just directly bbox intersect (with a horizontal
dilation) every left bracket with a right bracket. It's not smart enough to know
that "[i]" likely isn't a checkbox, but we can / should fix that with more image
searching to see if it's empty (something we could also borrow from box detect,
everything in the `get_checkboxes` that isn't in `get_boxes`.