quick and dirty patch
I removed all uses of SpaCy, hopefully, stepping around the conflict that presented in the Weaver with updates to other packages. I was only able to do this, however, by making use of the tools.suffolklitlab.org micro-services: `vectorize` and `passive`. You can see the testing results in `dev-testing/test package.ipynb` Unfortunately, this means that there are now a number of features that require a tools_token to work. I have not updated the documentation to make this clear, but I will open an issue. Also, I'm not sure if we want to push this change to PiPy or have the Weaver install from this branch as a hot fix. Obviously, where Weaver calls these functions no using tools, we'll need to pass a token.